refactor: separate api server and hide pm2 windows
