+++ zaphchat-frontend/index.html
@@ -3,6 +3,7 @@
 <head>
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <meta http-equiv="Content-Security-Policy" content="default-src 'self'; img-src 'self' data: https://picsum.photos; style-src 'self' 'unsafe-inline' https://fonts.googleapis.com; font-src https://fonts.gstatic.com; script-src 'self' 'unsafe-eval' https://cdn.tailwindcss.com https://esm.sh">
   <title>ZaphChat Dashboard</title>
   <script src="https://cdn.tailwindcss.com"></script>
   <style>
